Add files via upload
With a combination of both Excel and Power BI, I used Power Pivot to write queries that helped in cleaning of the dataset and ultimately coming up with visualizations to answer the questions of (1) which states had the highest orders? (2) Which gender had the highest orders based on the states? (3) Which year ultimately had the highest orders based on gender and state?